+++ src/h5core/h5u_readwrite.c
@@ -85,8 +85,9 @@ h5u_read_data (
 	const hid_t type
 	) {
         h5_file_p f = (h5_file_p)fh;
-	H5_CORE_API_ENTER (h5_err_t, "f=%p, name='%s', data=%p, type=%d",
-	                   f, name, data, type);
+	H5_CORE_API_ENTER (h5_err_t,
+			   "f=%p, name='%s', data=%p, type=%lld",
+	                   f, name, data, (long long int)type);
 	CHECK_TIMEGROUP (f);
 
 	struct h5u_fdata *u = f->u;
@@ -181,8 +182,9 @@ h5u_write_data (
 	const hid_t type	/*!< IN: Type of data */
 	) {
         h5_file_p f = (h5_file_p)fh;
-	H5_CORE_API_ENTER (h5_err_t, "f=%p, name='%s', data=%p, type=%d",
-	                   f, name, data, type);
+	H5_CORE_API_ENTER (h5_err_t,
+			   "f=%p, name='%s', data=%p, type=%lld",
+	                   f, name, data, (long long int)type);
 	CHECK_TIMEGROUP( f );
 	CHECK_WRITABLE_MODE( f );
